requirements.txt
anthropic>=0.25.0
google-cloud-texttospeech>=2.16.0
boto3>=1.34.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
python-dateutil>=2.9.0
pytz>=2024.1
pydub>=0.25.1
